+++ HaselTweaks/Tweaks/CastBarAetheryteNames.cs
@@ -21,7 +21,7 @@ public unsafe partial class CastBarAetheryteNames : ITweak
     private readonly IClientState _clientState;
     private readonly ExcelService _excelService;
     private readonly TextService _textService;
-    private readonly SeStringEvaluatorService _seStringEvaluator;
+    private readonly SeStringEvaluator _seStringEvaluator;
 
     private Hook<HaselActionManager.Delegates.OpenCastBar>? _openCastBarHook;
     private Hook<Telepo.Delegates.Teleport>? _teleportHook;
@@ -29,7 +29,6 @@ public unsafe partial class CastBarAetheryteNames : ITweak
     private TeleportInfo? _teleportInfo;
     private bool _isCastingTeleport;
 
-    public string InternalName => nameof(CastBarAetheryteNames);
     public TweakStatus Status { get; set; } = TweakStatus.Uninitialized;
 
     public void OnInitialize()