+++ HaselTweaks/Tweaks/EnhancedIsleworksAgenda.cs
@@ -1,11 +1,11 @@
 using Dalamud.Hooking;
 using Dalamud.Plugin.Services;
+using FFXIVClientStructs.FFXIV.Client.UI;
 using FFXIVClientStructs.FFXIV.Component.GUI;
 using HaselCommon.Services;
 using HaselTweaks.Config;
 using HaselTweaks.Enums;
 using HaselTweaks.Interfaces;
-using HaselTweaks.Structs;
 using HaselTweaks.Windows;
 
 namespace HaselTweaks.Tweaks;
@@ -21,7 +21,6 @@ public unsafe partial class EnhancedIsleworksAgenda : IConfigurableTweak
 
     private Hook<AddonMJICraftScheduleSetting.Delegates.ReceiveEvent>? _receiveEventHook;
 
-    public string InternalName => nameof(EnhancedIsleworksAgenda);
     public TweakStatus Status { get; set; } = TweakStatus.Uninitialized;
 
     public void OnInitialize()