+++ HaselTweaks/Tweaks/EnhancedMaterialList.cs
@@ -7,13 +7,14 @@ using Dalamud.Hooking;
 using Dalamud.Plugin.Services;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using FFXIVClientStructs.FFXIV.Client.Game.Character;
+using FFXIVClientStructs.FFXIV.Client.Game.UI;
+using FFXIVClientStructs.FFXIV.Client.UI;
 using FFXIVClientStructs.FFXIV.Client.UI.Agent;
 using FFXIVClientStructs.FFXIV.Component.GUI;
 using HaselCommon.Services;
 using HaselTweaks.Config;
 using HaselTweaks.Enums;
 using HaselTweaks.Interfaces;
-using HaselTweaks.Structs;
 using Lumina.Excel.Sheets;
 using Lumina.Extensions;
 using Lumina.Text;
@@ -35,14 +36,15 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
     private readonly IFramework _framework;
     private readonly IClientState _clientState;
     private readonly IGameInventory _gameInventory;
-    private readonly IAetheryteList _aetheryteList;
     private readonly AddonObserver _addonObserver;
     private readonly ExcelService _excelService;
     private readonly MapService _mapService;
     private readonly ItemService _itemService;
 
+    private delegate void AddonRecipeMaterialList_SetupRowDelegates(AddonRecipeMaterialList* thisPtr, nint a2, nint a3);
+
     private Hook<AgentRecipeMaterialList.Delegates.ReceiveEvent>? _agentRecipeMaterialListReceiveEventHook;
-    private Hook<AddonRecipeMaterialList.Delegates.SetupRow>? _addonRecipeMaterialListSetupRowHook;
+    private Hook<AddonRecipeMaterialList_SetupRowDelegates>? _addonRecipeMaterialListSetupRowHook;
     private Hook<AgentRecipeItemContext.Delegates.AddItemContextMenuEntries>? _addItemContextMenuEntriesHook;
 
     private bool _canRefreshMaterialList;
@@ -55,7 +57,6 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
     private DateTime _timeOfRecipeTreeRefresh;
     private bool _handleRecipeResultItemContextMenu;
 
-    public string InternalName => nameof(EnhancedMaterialList);
     public TweakStatus Status { get; set; } = TweakStatus.Uninitialized;
 
     public void OnInitialize()
@@ -64,8 +65,8 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
             AgentRecipeMaterialList.StaticVirtualTablePointer->ReceiveEvent,
             AgentRecipeMaterialListReceiveEventDetour);
 
-        _addonRecipeMaterialListSetupRowHook = _gameInteropProvider.HookFromAddress<AddonRecipeMaterialList.Delegates.SetupRow>(
-            AddonRecipeMaterialList.MemberFunctionPointers.SetupRow,
+        _addonRecipeMaterialListSetupRowHook = _gameInteropProvider.HookFromSignature<AddonRecipeMaterialList_SetupRowDelegates>(
+            "48 89 5C 24 ?? 48 89 54 24 ?? 48 89 4C 24 ?? 55 56 57 41 54 41 55 41 56 41 57 48 83 EC 50 49 8B 08",
             AddonRecipeMaterialListSetupRowDetour);
 
         _addItemContextMenuEntriesHook = _gameInteropProvider.HookFromAddress<AgentRecipeItemContext.Delegates.AddItemContextMenuEntries>(
@@ -181,7 +182,7 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
                 _canRefreshMaterialList = false;
                 return;
 
-            case AtkEventType.ListItemToggle:
+            case AtkEventType.ListItemClick:
                 if (!Config.ClickToOpenMap)
                     return;
 
@@ -192,11 +193,7 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
                 var rowData = **(nint**)(data + 0x08);
                 var itemId = *(uint*)(rowData + 0x04);
 
-                var itemRef = _excelService.CreateRef<Item>(itemId);
-                if (!itemRef.IsValid)
-                    return;
-
-                if (Config.DisableClickToOpenMapForCrystals && itemRef.Value.ItemUICategory.RowId == 59)
+                if (Config.DisableClickToOpenMapForCrystals && (!_excelService.TryGetRow<Item>(itemId, out var item) || item.ItemUICategory.RowId == 59))
                     return;
 
                 var tuple = GetPointForItem(itemId);
@@ -205,7 +202,7 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
 
                 var (totalPoints, point, cost, isSameZone, placeName) = tuple.Value;
 
-                _mapService.OpenMap(point, itemRef, "HaselTweaks"u8);
+                _mapService.OpenMap(point, itemId, "HaselTweaks"u8);
 
                 return;
 
@@ -411,11 +408,11 @@ public unsafe partial class EnhancedMaterialList : IConfigurableTweak
         {
             foreach (var p in gatheringPoints)
             {
-                foreach (var aetheryte in _aetheryteList)
+                foreach (var teleportInfo in Telepo.Instance()->TeleportList)
                 {
-                    if (aetheryte.AetheryteId == p!.TerritoryType.Value!.Aetheryte.RowId && (cost == 0 || aetheryte.GilCost < cost))
+                    if (teleportInfo.AetheryteId == p!.TerritoryType.Value!.Aetheryte.RowId && (cost == 0 || teleportInfo.GilCost < cost))
                     {
-                        cost = aetheryte.GilCost;
+                        cost = teleportInfo.GilCost;
                         point = p;
                         break;
                     }