+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [34.0.0] (2025-03-17)
+
+### New Tweak: Companion Color Preview
+
+Shows a small drop-down menu above the "Companion" window to select the color of your own chocobo. This is just to preview the colors and will not be saved!
+
+## [33.0.2] (2025-03-17)
+
+### Glamour Dresser Armoire Alert
+
+- **Fixed:** It was only possible to withdraw the first item due to an ImGui id conflict.
+
+## [33.0.1] (2025-03-17)
+
+### Glamour Dresser Armoire Alert
+
+- **Fixed:** The window wouldn't open when items were detected. My bad.
+
 ## [33.0.0] (2025-03-08)
 
 ### New Tweak: Enhanced Target Info
@@ -1941,6 +1959,9 @@ Refreshes the material list and recipe tree when you've crafted or gathered an i
 Opens duty finder for the duty you clicked on in the Wondrous Tails Journal.
 
 [unreleased]: https://github.com/Haselnussbomber/HaselTweaks/compare/main...dev
+[34.0.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v33.0.2...v34.0.0
+[33.0.2]: https://github.com/Haselnussbomber/HaselTweaks/compare/v33.0.1...v33.0.2
+[33.0.1]: https://github.com/Haselnussbomber/HaselTweaks/compare/v33.0.0...v33.0.1
 [33.0.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v32.1.1...v33.0.0
 [32.1.1]: https://github.com/Haselnussbomber/HaselTweaks/compare/v32.1.0...v32.1.1
 [32.1.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v32.0.4...v32.1.0